+++ CHANGELOG.md
@@ -29,6 +29,7 @@ The types of changes are:
 * Retry a DSR (FE) [#863](https://github.com/ethyca/fidesops/pull/938)
 * Add a Connection - Select a connector to configure (front end) [#760] (https://github.com/ethyca/fidesops/pull/987)
 * Add a Connection - Front End layout structure [#866] (https://github.com/ethyca/fidesops/pull/987)
+* Enable python function overrides for SaaS connector request execution [#986](https://github.com/ethyca/fidesops/pull/986)
 
 ### Changed